Move out types from requests to walletConnect
