--- src/redux/__mocks__/requests.ts
@@ -1,3 +1,0 @@
-import { jest } from '@jest/globals';
-
-export const REQUESTS_UPDATE_REQUESTS_TO_APPROVE = 'requests/REQUESTS_UPDATE_REQUESTS_TO_APPROVE';
