+++ src/resources/transactions/transactionSimulation.ts
@@ -4,7 +4,7 @@ import { RainbowError, logger } from '@/logger';
 import { metadataPOSTClient } from '@/graphql';
 import { TransactionErrorType, TransactionScanResultType, TransactionSimulationResult } from '@/graphql/__generated__/metadataPOST';
 import { isNil } from 'lodash';
-import { RequestData } from '@/redux/requests';
+import { RequestData } from '@/walletConnect/types';
 import { ChainId } from '@/chains/types';
 
 type SimulationArgs = {
